feat(ui): change recovery seed phrase to shared component
