+++ src/ui/components/RecoverySeedPhraseModule/RecoverySeedPhraseModule.types.ts
@@ -1,0 +1,21 @@
+interface SeedPhraseInfo {
+  value: string;
+  suggestions: string[];
+}
+
+interface RecoverySeedPhraseModuleProps {
+  title?: string;
+  description: string;
+  testId: string;
+  onVerifySuccess: () => void;
+}
+
+interface RecoverySeedPhraseModuleRef {
+  clearState: () => void;
+}
+
+export type {
+  SeedPhraseInfo,
+  RecoverySeedPhraseModuleProps,
+  RecoverySeedPhraseModuleRef,
+};
